# specs/001-web-ui-api/contracts/api-spec.yaml
openapi: 3.1.0
info:
  title: Ginkgo API Server
  description: Ginkgo量化交易系统API接口
  version: 1.0.0
  contact:
    name: Ginkgo Team
    email: support@ginkgo.trade
servers:
  - url: http://localhost:8000
    description: 开发环境
  - url: https://api.ginkgo.trade
    description: 生产环境
...
paths:
...
  /api/portfolio:
    get:
      tags: [portfolio]
      summary: 获取Portfolio列表
      description: 获取当前用户的所有Portfolio
      security:
        - BearerAuth: []
      parameters:
        - name: mode
          in: query
          schema:
            type: string
            enum: [BACKTEST, PAPER, LIVE]
          description: 按运行模式筛选
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PortfolioSummary'
    post:
      tags: [portfolio]
      summary: 创建Portfolio
      description: 创建新的Portfolio
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PortfolioCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioResponse'
  /api/portfolio/{uuid}:
    get:
      tags: [portfolio]
      summary: 获取Portfolio详情
      security:
        - BearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioDetail'
        '404':
          description: Portfolio不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/arena/portfolios:
    get:
      tags: [arena]
      summary: 获取竞技场Portfolio列表
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ArenaPortfolio'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: integer
          example: 401
        message:
          type: string
          example: "Unauthorized"
        details:
          type: object
          additionalProperties: true
...
    PortfolioCreate:
      type: object
      required:
        - name
        - initial_cash
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        initial_cash:
          type: number
          format: float
          minimum: 0
          example: 100000
        risk_config:
          type: object
          properties:
            stop_loss:
              type: number
              format: float
            take_profit:
              type: number
              format: float
    PortfolioSummary:
      type: object
      properties:
        uuid:
          type: string
        name:
          type: string
        mode:
          type: string
          enum: [BACKTEST, PAPER, LIVE]
        state:
          type: string
          enum: [INITIALIZED, RUNNING, PAUSED, STOPPED]
        config_locked:
          type: boolean
        net_value:
          type: number
          format: float
        created_at:
          type: string
          format: date-time
    PortfolioDetail:
      allOf:
        - $ref: '#/components/schemas/PortfolioSummary'
        - type: object
          properties:
            initial_cash:
              type: number
              format: float
            current_cash:
              type: number
              format: float
            positions:
              type: array
              items:
                type: object
                properties:
                  code:
                    type: string
                  volume:
                    type: integer
                  cost_price:
                    type: number
                    format: float
                  current_price:
                    type: number
                    format: float
            strategies:
              type: array
              items:
                $ref: '#/components/schemas/StrategySummary'
            risk_alerts:
              type: array
              items:
                $ref: '#/components/schemas/RiskAlertSummary'
    StrategySummary:
      type: object
      properties:
        uuid:
          type: string
        name:
          type: string
        type:
          type: string
        weight:
          type: number
          format: float
        performance:
          type: object
          properties:
            return:
              type: number
              format: float
            sharpe:
              type: number
              format: float
            max_drawdown:
              type: number
              format: float
...
    ArenaPortfolio:
      type: object
      properties:
        uuid:
          type: string
        name:
          type: string
        return:
          type: number
          format: float
        sharpe:
          type: number
          format: float
        max_drawdown:
          type: number
          format: float
        color:
          type: string
          example: "#1890ff"
...
    RiskAlertSummary:
      type: object
      properties:
        uuid:
          type: string
        type:
          type: string
          enum: [STOP_LOSS, TAKE_PROFIT, POSITION_LIMIT]
        level:
          type: string
          enum: [INFO, WARNING, ERROR, CRITICAL]
        message:
          type: string
        triggered_at:
          type: string
          format: date-time
        handled:
          type: boolean